check jwt on certaing Usuario requests

+++ api-usuario/src/main/java/br/edu/utfpr/api_usuario/controllers/UsuarioController.java
@@ -7,6 +7,7 @@ import br.edu.utfpr.api_usuario.repositories.UsuarioRepository;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.oauth2.server.resource.authentication.JwtAuthenticationToken;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -29,11 +30,15 @@ public class UsuarioController {
 	}
 
 	@GetMapping("/{id}")
-	public ResponseEntity<Usuario> getById(@PathVariable(name = "id") Long id) {
+	public ResponseEntity<Usuario> getById(@PathVariable(name = "id") Long id, JwtAuthenticationToken jwt) {
+		String loggedEmail = jwt.getToken().getSubject();
+
 		Usuario usuario = this.repository.findById(id).orElse(null);
 
 		if (usuario == null) {
 			return ResponseEntity.status(HttpStatus.NOT_FOUND).body(null);
+		} else if (!usuario.getEmail().equals(loggedEmail)) {
+			return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(null);
 		} else {
 			return ResponseEntity.ok().body(usuario);
 		}
@@ -51,10 +56,12 @@ public class UsuarioController {
 	}
 
 	@PutMapping("/{id}")
-	public ResponseEntity<String> updateOne(@PathVariable(name = "id") Long id, @RequestBody UsuarioDTO usuarioDTO) {
+	public ResponseEntity<String> updateOne(@PathVariable(name = "id") Long id, @RequestBody UsuarioDTO usuarioDTO, JwtAuthenticationToken jwt) {
 		Usuario usuario = this.repository.findById(id).orElse(null);
 
-		if (usuario != null) {
+		String loggedEmail = jwt.getToken().getSubject();
+
+		if (usuario != null && usuario.getEmail().equals(loggedEmail)) {
 			usuario.setNome(usuarioDTO.nome());
 			usuario.setEmail(usuarioDTO.email());
 
@@ -67,9 +74,12 @@ public class UsuarioController {
 	}
 
 	@DeleteMapping(path = "/{id}")
-	public ResponseEntity<String> deleteOne(@PathVariable(name = "id") Long id) {
+	public ResponseEntity<String> deleteOne(@PathVariable(name = "id") Long id, JwtAuthenticationToken jwt) {
 		Usuario usuario = this.repository.findById(id).orElse(null);
-		if (usuario != null) {
+
+		String loggedEmail = jwt.getToken().getSubject();
+
+		if (usuario != null && loggedEmail.equals(usuario.getEmail())) {
 			this.repository.delete(usuario);
 			return ResponseEntity.status(HttpStatus.OK).body("Usuario deletado com sucesso!");
 		} else {
